silo-core: consistency for lt check

+++ silo-vaults/contracts/incentives/claiming-logics/SiloIncentivesControllerCL.sol
@@ -35,7 +35,7 @@ contract SiloIncentivesControllerCL is IIncentivesClaimingLogic {
 
             VAULT_INCENTIVES_CONTROLLER.immediateDistribution(
                 accruedRewards[i].rewardToken,
-                uint104(accruedRewards[i].amount)
+                accruedRewards[i].amount
             );
         }
     }
